requirements.txt
python-dotenv==1.0.1
aiogram==2.23.1
pyrogram==2.0.106
tgcrypto==1.2.5